Minor Revision
To be sure, I don't know if Indonesian vaccination rate is daily update (based on End of Day tabulation) or result from the previous working day. If it comes from the previous working day, Sinopharm should start to be counted on 19 May, and Moderna should start to be counted from 17 July. Since both vaccination is only started to be used in the previous day.